Add logout option in NavBar
